+++ core/src/plugins/kubernetes/kubectl.ts
@@ -17,9 +17,8 @@ import { PluginToolSpec } from "../../plugin/tools"
 import { PluginContext } from "../../plugin-context"
 import { KubeApi, KubernetesError } from "./api"
 import { pathExists } from "fs-extra"
-import { ConfigurationError } from "../../exceptions"
+import { ChildProcessError, ConfigurationError } from "../../exceptions"
 import { requestWithRetry, RetryOpts } from "./retry"
-import { ChildProcessError } from "../../util/util"
 
 // Corresponds to the default prune whitelist in `kubectl`.
 // See: https://github.com/kubernetes/kubectl/blob/master/pkg/cmd/apply/prune.go#L176-L192
