+++ core/src/plugins/kubernetes/local/minikube.ts
@@ -6,7 +6,8 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-import { ChildProcessError, exec } from "../../../util/util"
+import { ChildProcessError } from "../../../exceptions"
+import { exec } from "../../../util/util"
 
 /**
  * Automatically set docker environment variables for minikube
