+++ core/test/unit/src/util/util.ts
@@ -17,93 +17,19 @@ import {
   spawn,
   relationshipClasses,
   isValidDateInstance,
-  ChildProcessError,
 } from "../../../../src/util/util"
 import { expectError } from "../../../helpers"
 import { splitLast, splitFirst } from "../../../../src/util/string"
 import { getRootLogger } from "../../../../src/logger/logger"
 import { dedent } from "../../../../src/util/string"
 import { safeDumpYaml } from "../../../../src/util/serialization"
+import { ChildProcessError } from "../../../../src/exceptions"
 
 function isLinuxOrDarwin() {
   return process.platform === "darwin" || process.platform === "linux"
 }
 
 describe("util", () => {
-  describe("ChildProcessError", () => {
-    it("formats an appropriate error message", () => {
-      const err = new ChildProcessError({
-        code: 1,
-        cmd: "ls",
-        args: ["some-dir"],
-        stderr: "dir not found",
-        stdout: "",
-        output: "dir not found",
-      })
-      expect(err.message).to.equal(dedent`
-        Command "ls some-dir" failed with code 1:
-
-        dir not found
-      `)
-    })
-    it("should ignore emtpy args", () => {
-      const err = new ChildProcessError({
-        code: 1,
-        cmd: "ls",
-        args: [],
-        stderr: "dir not found",
-        stdout: "",
-        output: "dir not found",
-      })
-      expect(err.message).to.equal(dedent`
-        Command "ls" failed with code 1:
-
-        dir not found
-      `)
-    })
-    it("should include output if it's not the same as the error", () => {
-      const err = new ChildProcessError({
-        code: 1,
-        cmd: "ls some-dir",
-        args: [],
-        stderr: "dir not found",
-        stdout: " and some more output",
-        output: "dir not found and some more output",
-      })
-      expect(err.message).to.equal(dedent`
-        Command "ls some-dir" failed with code 1:
-
-        dir not found
-
-        Here's the full output:
-
-        dir not found and some more output
-      `)
-    })
-    it("should include the last 100 lines of output if output is very long", () => {
-      const output = "All work and no play\n"
-      const outputFull = output.repeat(102)
-      const outputPartial = output.repeat(99) // This makes 100 lines in total
-
-      const err = new ChildProcessError({
-        code: 1,
-        cmd: "ls some-dir",
-        args: [],
-        stderr: "dir not found",
-        stdout: outputFull,
-        output: outputFull,
-      })
-      expect(err.message).to.equal(dedent`
-        Command "ls some-dir" failed with code 1:
-
-        dir not found
-
-        Here are the last 100 lines of the output:
-
-        ${outputPartial}
-      `)
-    })
-  })
   describe("exec", () => {
     before(function () {
       // These tests depend the underlying OS and are only executed on macOS and linux
